Abstracted methods to other files and did some refactoring
