+++ app.log
@@ -1,3 +1,5 @@
+Start of session
+2021-08-11-19:07:48
 Building Migrations
 Opening {db}
 adding Adam Page
@@ -14144,3 +14146,73 @@ processing NJPW
 processing NXT
 processing ROH
 processing WWE
+Matches = 9
+Building Show
+49 was rolled with 1 and 101 indexes
+Match Roll for index 1 is 49
+53 was rolled with 1 and 101 indexes
+Gender Picker is below
+male
+Singles Match gendered
+9 was rolled with 1 and 101 indexes
+Match Roll for index 2 is 9
+1 was rolled with 1 and 101 indexes
+Combatent Picker
+One on One
+24/7 match
+87 was rolled with 1 and 101 indexes
+Match Roll for index 3 is 87
+Tag Match Maker
+71 was rolled with 1 and 101 indexes
+Gender Picker is below
+female
+17 was rolled with 1 and 24 indexes
+21 was rolled with 1 and 24 indexes
+Tag Match
+91 was rolled with 1 and 101 indexes
+Match Roll for index 4 is 91
+Tag Match Maker
+55 was rolled with 1 and 101 indexes
+Gender Picker is below
+male
+3 was rolled with 1 and 24 indexes
+5 was rolled with 1 and 24 indexes
+Tag Match
+14 was rolled with 1 and 101 indexes
+Match Roll for index 5 is 14
+97 was rolled with 1 and 101 indexes
+Combatent Picker
+Fatal Four Way
+24/7 match
+51 was rolled with 1 and 101 indexes
+Match Roll for index 6 is 51
+Team Sizer
+43 was rolled with 1 and 101 indexes
+74 was rolled with 1 and 101 indexes
+Gender Picker is below
+female
+Team size = 6-Woman
+X man/woman Tag Match
+100 was rolled with 1 and 101 indexes
+Match Roll for index 7 is 100
+1 was rolled with 1 and 6 indexes
+5 was rolled with 2 and 6 indexes
+1 on 5 Handicap match
+1 was rolled with 1 and 101 indexes
+Match Roll for index 8 is 1
+27 was rolled with 1 and 101 indexes
+Combatent Picker
+Triangle Match
+24/7 match
+76 was rolled with 1 and 101 indexes
+Main even roll = 76
+35 was rolled with 1 and 101 indexes
+Combatent Picker
+Triangle Match
+Title is Raw Tag Championship
+#1 contender match
+91 was rolled with 1 and 101 indexes
+Combatent Picker
+Fatal Four Way
+Title is US Championship
+Title match
